requirements.txt
crewai[tools]
python-dotenv
google-genai
Pillow
